Fixing the rest of the tests
