--- tests/library/child/conftest.py
@@ -1,6 +1,0 @@
-from pytest_bdd import given
-
-
-@given("I have the overriden fixture")
-def overridable():
-    return "child"
